Changing from a module/activity to a local plugin
This module only existed as a placeholder to load in database changes
for the Mahara view submissions assignment type, because the assignment
type subplugins aren't allowed to have their own db files. In Moodle 2 it
makes more sense for this to be a local plugin (since you can have multiple
local plugins).